added functionality for non existent entry choice

+++ application/routes.py
@@ -14,22 +14,28 @@ def read():
     all_tasks = Todo.query.all()
     tasks_string = ""
     for task in all_tasks:
-        tasks_string += "<br>"+ task.name+ "- " +task.status
+        tasks_string += "<br>"+ str(task.id) + ". " + task.name+ "- " +task.status
     return tasks_string
 
 @app.route('/update/<int:number>/<name>')
 def update(name, number):
     task = Todo.query.filter_by(id=number).first()
-    task.name = name
-    db.session.commit()
-    return task.name
+    if task is not None:
+        task.name = name
+        db.session.commit()
+        return task.name
+    else:
+        return "entry does not exist"
 
 @app.route('/delete/<int:number>')
 def delete(number):
     task = Todo.query.filter_by(id=number).first()
-    db.session.delete(task)
-    db.session.commit()
-    return task.name
+    if task is not None:
+        db.session.delete(task)
+        db.session.commit()
+        return task.name
+    else:
+        return "Entry does not exist"
 
 @app.route('/count')
 def count():
@@ -40,6 +46,9 @@ def count():
 @app.route('/status/<int:number>/<name>')
 def status(number, name):
     task = Todo.query.filter_by(id=number).first()
-    task.status = name
-    db.session.commit()
-    return task.status
+    if task is not None:
+        task.status = name
+        db.session.commit()
+        return task.status
+    else:
+        return "Entry does not exist"
